+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/web/endpoints/FlagFromTagDimensionDataServletSpec.groovy
@@ -198,7 +198,6 @@ class FlagFromTagDimensionDataServletSpec extends BaseDataServletComponentSpec {
                 "type" : "table"
             },
             "dimensions": [
-                "filteringDimension",
                {
                   "dimension":"breed",
                   "outputName":"flagFromTag",