+++ CHANGELOG.md
@@ -11,7 +11,7 @@ Current
 
 - [Fix: Add withDimension in TemplateDruidQuery](https://github.com/yahoo/fili/pull/1058)
     * Modify a TDQ to add a dependent dimension.
-    * Added getAllReferencedDimension in DataApiRequest to get request,metric,filter dimensions and pass it along to query builder.
+    * Added getAllGroupingDimensions in DataApiRequest to get request,metric,filter dimensions and pass it along to query builder.
 
 - [Fix: Antlr Sort Parser didn't work.  Wasn't tested.  Patched and Tested](https://github.com/yahoo/fili/issues/1050)
     * Moved binding out of `Sorts.g4` grammar and into `SortDirection`.